Add a logging handler
This sets up logging, sending logs to stderr, with a helpful format that
includes the file and function generating the logs.  It is set up such
that `import *` will import functions that log to the appropriate level.

NOTE: Right now, all logs (DEBUG level and above) are being output.
Later on, this default should be changed to WARNING, and an environment
variable (or command-line param) queried to set a different log level.