The start of the database code
To keep track of what needs to be done, we'll use a SQLite DB.  This
code uses aiosqlite, since a cunk of the program will be async.

This code gets the database path, and handles opening the database,
verifying the DB is one of ours (by checking application ID), and doing
any needed schema upgrades.

NOTE: Non-async code needs to handle opening the DB itself, using the
built-in sqlite3 module.

TODO: Check that the DB schema isn't newer than what we expect.