The start of the Globus code
This is where all of the Globus code will be.  It currently handles
token storage (using the DB), and checking for valid tokens.  It's up to
calling code to trigger a login, if needed.

TODO: Request Globus Transfer scopes.

Next, code will be added to fetch a list of transfers, fetch info on a
single transfer, and fetch a list of transferred files (for a completed
transfer).