The start of the CLI code
This is what gets run when `aws-copy-pt` is run.  It starts by checking
(and initializing, if needed) the DB.  Then it checks for valid Globus
credentials, running the user through a Globus login (if needed).  This
login will check if the user has a GUI, and will fall back to manual
code-entry if needed.

The CLI will also check for a non-interactive run, and exit (for now).