globus: Add collection search and retrieval
This adds code needed to search for, and validate, Globus collections.

Searching will search for the 10 most recently-used collections (where
"recently-used" is defined by Globus).  Validation will take a
collection UUID as input.

In all cases, the returned value is a dataclass (or list of dataclasses)
consisting of the UUID, the display name, and the server FQDN.  The
server FQDN will be needed later, to do authorization and grab
credentials.