cli: Add code to select Globus collection
This starts out the code to get a Globus collection.  It searches for
recently-used collections.  If any are found, collection names and UUIDs
are presented to the user for selection.

The user is also able to enter a collection UUID, to account for
collections not on the list.  The code checks the response, either that
it's a valid list index, or that it's a valid collection.

Next up will be to update the list of Globus scopes, validate the
collection is an S3 collection, and store the collection info in the DB.